refactor PureKerasModel.predict test
